rename: fix spelling in (Renders? Renderers?) Choosing Renderers for now
